Move loadTestableCollectionFromDocsAsync to backendRequests.ts
