+++ teachertool/src/transforms/loadCatalogAsync.ts
@@ -1,7 +1,7 @@
+import { loadTestableCollectionFromDocsAsync } from "../services/backendRequests";
 import { stateAndDispatch } from "../state";
 import * as Actions from "../state/actions";
 import { CatalogCriteria } from "../types/criteria";
-import { loadTestableCollectionFromFileAsync } from "../utils/fileSystemHelpers";
 
 const prodFiles = [
     "/teachertool/catalog.json", // target-specific catalog
@@ -10,6 +10,6 @@ const prodFiles = [
 
 export async function loadCatalogAsync() {
     const { dispatch } = stateAndDispatch();
-    const fullCatalog = await loadTestableCollectionFromFileAsync<CatalogCriteria>(prodFiles, "criteria");
+    const fullCatalog = await loadTestableCollectionFromDocsAsync<CatalogCriteria>(prodFiles, "criteria");
     dispatch(Actions.setCatalog(fullCatalog));
 }
